requirements.txt
fastapi
uvicorn
pydantic
python-docx
striprtf
python-pptx
trafilatura
pymupdf4llm
pypdf
google-cloud-storage
httpx
